To fix the rate limiting for /healthz, reorganize the app/__init__ a bunch to support decorators
